pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claude-code-discord-bridge"
version = "0.1.0"
description = "Discord frontend for Claude Code CLI - chat with Claude Code via Discord threads"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "discord.py>=2.4,<3.0",
    "aiosqlite>=0.20,<1.0",
    "python-dotenv>=1.0,<2.0",
]

[project.optional-dependencies]
api = ["aiohttp>=3.9"]

[tool.hatch.build.targets.wheel]
packages = ["claude_discord"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "aiohttp>=3.9",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]
